refactor: move url to llrt_modules
