+++ llrt_core/Cargo.toml
@@ -78,7 +78,6 @@ flate2 = { version = "1.0.30", features = [
 ], default-features = false }
 brotlic = "0.8.2"
 rustls-pemfile = "2.1.2"
-url = "=2.5.1"
 
 [build-dependencies]
 rquickjs = { version = "0.6.2", features = [
