+++ llrt_core/src/modules/mod.rs
@@ -1,9 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-
 pub use llrt_modules::{
     buffer, child_process, crypto, exceptions, fs, navigator, net, os, path, perf_hooks, process,
-    zlib,
+    url, zlib,
 };
 
 pub mod console;
@@ -12,5 +11,4 @@ pub mod events;
 pub mod http;
 pub mod llrt;
 pub mod module;
-pub mod url;
 pub mod util;
