+++ llrt_core/src/utils/mod.rs
@@ -1,10 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-pub mod class;
 pub mod clone;
 pub mod io;
 pub mod mc_oneshot;
-pub mod object;
 pub mod result {
     pub use llrt_utils::result::*;
 }
