+++ llrt_modules/src/modules/url/url_class.rs
@@ -1,14 +1,14 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
+use std::{cell::RefCell, rc::Rc};
 
-use super::url_search_params::URLSearchParams;
-use crate::utils::result::ResultExt;
 use rquickjs::{
     atom::PredefinedAtom, class::Trace, function::Opt, Class, Coerced, Ctx, Exception, FromJs,
     IntoJs, Null, Object, Result, Value,
 };
-use std::{cell::RefCell, path::PathBuf, rc::Rc, str::FromStr};
-use url::{quirks, Url};
+use url_crate::{quirks, Url};
+
+use super::url_search_params::URLSearchParams;
 
 /// Naively checks for hostname delimiter, a colon ":", that's *probably* not
 /// part of an IPv6 address
@@ -301,100 +301,4 @@ pub fn url_to_http_options<'js>(ctx: Ctx<'js>, url: Class<'js, URL<'js>>) -> Res
     }
 
     Ok(obj)
-}
-
-pub fn domain_to_unicode(domain: &str) -> String {
-    quirks::domain_to_unicode(domain)
-}
-
-pub fn domain_to_ascii(domain: &str) -> String {
-    quirks::domain_to_ascii(domain)
-}
-
-//options are ignored, no windows support yet
-pub fn path_to_file_url<'js>(ctx: Ctx<'js>, path: String, _: Opt<Value>) -> Result<URL<'js>> {
-    let url = Url::from_file_path(&path)
-        .map_err(|_| Exception::throw_type(&ctx, &["Path is not absolute: ", &path].concat()))?;
-
-    URL::from_url(ctx, url)
-}
-
-//options are ignored, no windows support yet
-pub fn file_url_to_path<'js>(ctx: Ctx<'js>, url: Value<'js>) -> Result<String> {
-    let url_string = if let Ok(url) = Class::<URL>::from_value(&url) {
-        url.borrow().to_string()
-    } else {
-        url.get::<Coerced<String>>()?.to_string()
-    };
-
-    let path = if let Some(path) = &url_string.strip_prefix("file://") {
-        path.to_string()
-    } else {
-        url_string
-    };
-
-    Ok(PathBuf::from_str(&path)
-        .or_throw(&ctx)?
-        .to_string_lossy()
-        .to_string())
-}
-
-pub fn url_format<'js>(url: Class<'js, URL<'js>>, options: Opt<Value<'js>>) -> Result<String> {
-    let url = url.borrow();
-    let mut string = url.protocol();
-    string.push_str("//");
-
-    let mut include_fragment = true;
-    let mut unicode_encode = false;
-    let mut include_auth = true;
-    let mut include_search = true;
-
-    // Parse options if provided
-    if let Some(options) = options.into_inner() {
-        if let Some(options) = options.as_object() {
-            if let Ok(value) = options.get("unicode") {
-                unicode_encode = value;
-            }
-            if let Ok(value) = options.get("auth") {
-                include_auth = value;
-            }
-            if let Ok(value) = options.get("fragment") {
-                include_fragment = value;
-            }
-            if let Ok(value) = options.get("search") {
-                include_search = value
-            }
-        }
-    }
-
-    if include_auth {
-        let username = url.username();
-        let password = url.password();
-        if !username.is_empty() {
-            string.push_str(&username);
-            if !password.is_empty() {
-                string.push(':');
-                string.push_str(&password);
-            }
-            string.push('@');
-        }
-    }
-
-    if unicode_encode {
-        string.push_str(&domain_to_unicode(&url.host()));
-    } else {
-        string.push_str(&url.host());
-    }
-
-    string.push_str(&url.pathname());
-
-    if include_search {
-        string.push_str(&url.search());
-    }
-
-    if include_fragment {
-        string.push_str(&url.hash());
-    }
-
-    Ok(string)
 }
