+++ llrt_utils/src/lib.rs
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 #![cfg_attr(rust_nightly, feature(array_chunks))]
 pub mod bytes;
-pub mod error_messages;
-
+pub mod class;
 #[cfg(feature = "ctx")]
 pub mod ctx;
 #[cfg(feature = "encoding")]
 pub mod encoding;
 pub mod error;
+pub mod error_messages;
 #[cfg(feature = "fs")]
 pub mod fs;
 pub mod module;
